fixup image tests
